Trinamic seems to be working on 24V with these settings

+++ Marlin/stepper.h
@@ -80,7 +80,7 @@ void quickStop();
   int trinamicGetStalled(int axis);
   int trinamicGetStallGuard(int axis);
   uint32_t trinamicGetDRVSTATUS(int axis);
-  int trinamicGetGStat(int axis);
+  uint8_t trinamicGetGStat(int axis);
